Fix(fontend/withdrawal): credit can now be float number

+++ frontend/src/pages/WithdrawalInfo.tsx
@@ -74,7 +74,7 @@ const WithdrawInfoPage = () => {
     const status = tmp ? parseInt(tmp) : tmp;
 
     tmp = creditRef.current?.value;
-    const credit = tmp ? parseInt(tmp) : tmp;
+    const credit = tmp ? parseFloat(tmp) : tmp;
 
     const course1 = course1Ref.current?.value;
     const course2 = course2Ref.current?.value;
